+++ packages/permissionless/actions/smartAccount/prepareUserOperationErc20.ts
@@ -29,8 +29,8 @@ import {
 import { estimateFeesPerGas } from "viem/actions"
 import { AccountNotFoundError } from "viem/errors/account"
 import { getAction, parseAccount } from "viem/utils"
-import { getRequiredPrefund } from "../utils/getRequiredPrefund"
-import { getTokenQuotes } from "./pimlico"
+import { getRequiredPrefund } from "../../utils/getRequiredPrefund"
+import { getTokenQuotes } from "../pimlico"
 
 const defaultParameters = [
     "factory",
